moving classes to new constants package
