+++ phenotools-builder/src/main/java/org/phenopackets/phenotools/builder/builders/constants/Onset.java
@@ -1,4 +1,4 @@
-package org.phenopackets.phenotools.builder.builders;
+package org.phenopackets.phenotools.builder.builders.constants;
 
 import org.phenopackets.schema.v2.core.OntologyClass;
