+++ phenotools-cli/src/main/java/org/phenopackets/phenotools/examples/Thrombocytopenia2.java
@@ -2,6 +2,7 @@ package org.phenopackets.phenotools.examples;
 
 import org.phenopackets.phenotools.builder.PhenopacketBuilder;
 import org.phenopackets.phenotools.builder.builders.*;
+import org.phenopackets.phenotools.builder.builders.constants.Status;
 import org.phenopackets.schema.v2.Phenopacket;
 
 import static org.phenopackets.phenotools.builder.builders.OntologyClassBuilder.ontologyClass;
